update: OGP and meta
