+++ src/styles/app/contact/contact.module.scss
@@ -64,7 +64,7 @@
     .language_switching_box {
       position: sticky;
       user-select: none;
-      bottom: .5rem;
+      bottom: 1.5rem;
       width: 100%;
       height: 2.5rem;
       left: 0;
