+++ src/styles/app/page.module.scss
@@ -7,7 +7,7 @@
 
   .language_switching_box {
     position: sticky;
-    bottom: 0.5rem;
+    bottom: 1.5rem;
     width: 100%;
     height: 2.5rem;
     left: 0;
